New Wiener-Khinchin-based DDM calculation.
Started working on new type of DDM algorithm which is ~10 times faster
than the previous one, thanks to fast temporal FFT and Wiener Khinchin
theorem. Need to add option to choose the type of algorithm and test
more (noticable artefact for zero frequency).

+++ src/includes/stack.cpp
@@ -86,14 +86,14 @@ std::runtime_error Stack::error_reading(const std::string& msg) {
 
 // === ENCODING SPECIFIC === //
 void Stack::load_M12P_images(int N) {
-    for(int i = 0; i < N; ++i) {
+    for(long i = 0; i < N; ++i) {
         this->load_next_M12P_frame(i * this->image_size);
     }
 }
 
 
 // memvar offset
-void Stack::load_next_M12P_frame(int offset) {
+void Stack::load_next_M12P_frame(long offset) {
     /*
      * Load next image into the buffer ;
      * Performs sanity check to verify we're reading the right bytes
